family_budget/requirements.txt
black==23.9.1
Django==4.2.5
django-cors-headers==4.2.0
djangorestframework==3.14.0
flake8==6.1.0
isort==5.12.0
python-dotenv==1.0.0
django-filter==23.3
